[ALPS08276295] cmdq: write_addr_high data type change to u32
write_addr_high data type change to u32

MTK-Commit-Id: 5b8db61ac81705b0f70ef64c206f7cddd9c3dc3e

CR-Id: ALPS08276295
Feature: cmdq_gce

+++ include/linux/mailbox/mtk-cmdq-mailbox-ext.h
@@ -197,7 +197,7 @@ struct cmdq_pkt {
 	struct cmdq_append append;
 
 	struct work_struct	destroy_work;
-	u16			write_addr_high;
+	u32			write_addr_high;
 	struct device	*share_dev;
 	size_t			create_instr_cnt;
 	bool			timeout_dump_hw_trace;
